fixes for static builds, seems github version is old

+++ src/os/windows/client/IOverlap.cpp
@@ -24,7 +24,7 @@ IOverlap::IOverlap(std::shared_ptr<Socket> sckt, size_t bufferSize)
     mBuffer = std::shared_ptr<char>(new char[bufferSize]);
     memset(mBuffer.get(), 0, bufferSize);
     mWsaBuffer.buf = mBuffer.get();
-    mWsaBuffer.len = bufferSize;
+    mWsaBuffer.len = (ULONG)bufferSize;
 }
 
 //------------------------------------------------------------------------------
@@ -39,7 +39,7 @@ IOverlap::IOverlap(std::shared_ptr<Socket> sckt, std::shared_ptr<utilities::Byte
     mBuffer = std::shared_ptr<char>(new char[stream->size()]);
     memcpy(mBuffer.get(), stream->buffer(), stream->size());
     mWsaBuffer.buf = mBuffer.get();
-    mWsaBuffer.len = stream->size();
+    mWsaBuffer.len = (ULONG)stream->size();
 }
 
 //------------------------------------------------------------------------------
